test fix and addressed gitar comment
