+++ openmetadata-ui/src/main/resources/ui/playwright/e2e/Features/ContextCenterPermission.spec.ts
@@ -13,9 +13,9 @@
 
 import {
   APIRequestContext,
+  test as base,
   expect,
   Page,
-  test as base,
 } from '@playwright/test';
 import { KnowledgeCenterClass } from '../../support/entity/KnowledgeCenterClass';
 import { UserClass } from '../../support/user/UserClass';
@@ -413,7 +413,7 @@ test.describe('Context Center Permissions', () => {
       deleteAllOwnMemoryId,
       allPermissionOwnMemoryId,
       viewOnlyOwnMemoryId,
-      // earlyAlphabetMemoryId,
+      earlyAlphabetMemoryId,
     ]) {
       if (memoryId) {
         await apiContext
@@ -427,7 +427,7 @@ test.describe('Context Center Permissions', () => {
       editAllUser,
       deleteAllUser,
       allPermissionUser,
-      // earlyAlphabetUser,
+      earlyAlphabetUser,
     ]) {
       if (user?.responseData?.id) {
         await user.delete(apiContext);
